Adjust container spacing and font sizes
